vizro_chat/pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
dependencies = [
  "vizro>=0.1.32",
  "dash-bootstrap-components>=1.0.0",
  "python-dotenv>=0.21.0",
  "dash>=2.0.0"
]
description = "A chat component for Vizro dashboards"
name = "vizro-chat"
requires-python = ">=3.9"
version = "0.0.1"

[project.optional-dependencies]
dev = [
  "pytest>=7.0"
]
openai = [
  "openai>=1.0.0"
]

[tool.hatch.build]
include = [
  "vizro_chat/**/*.py"
]

[tool.hatch.build.targets.wheel]
packages = ["vizro_chat"]
